backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
# innovation-ia==0.1.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
gunicorn==23.0.0
uvicorn[standard]==0.34.0
sendgrid==6.11.0
twilio==9.2.2
email-validator
google-generativeai
httpx
slowapi
python-multipart
stripe
pytest
mercadopago
jinja2
redis>=5.0.0
celery>=5.3.0
python-multipart
# AI and Workers
arq>=0.25.0
